+++ src/surrealist/ql/statements/__init__.py
@@ -1,0 +1,10 @@
+from .create import Create
+from .delete import Delete
+from .insert import Insert
+from .live import Live
+from .remove import Remove
+from .select import Select
+from .show import Show
+from .update import Update
+
+__all__ = ("Create", "Delete", "Insert", "Live", "Remove", "Select", "Show", "Update")
