+++ src/surrealist/ql/statements/delete.py
@@ -1,22 +1,18 @@
 from typing import List, Optional
 
 from surrealist import Connection
-from surrealist.ql.create_statements import CreateUseSetContent
-from surrealist.ql.statement import Statement
+from .common_statements import CanUseWhere
+from .statement import Statement
 from surrealist.utils import OK
 
 
-class Create(Statement, CreateUseSetContent):
+class Delete(Statement, CanUseWhere):
 
-    def __init__(self, connection: Connection, table_name: str, record_id: Optional[str] = None):
+    def __init__(self, connection: Connection, table_name: str, record_id:Optional[str]=None):
         super().__init__(connection)
         self._table_name = table_name
         self._only = False
         self._record_id = record_id
-
-    def only(self) -> "Create":
-        self._only = True
-        return self
 
     def validate(self) -> List[str]:
         return [OK]
@@ -24,4 +20,8 @@ class Create(Statement, CreateUseSetContent):
     def _clean_str(self):
         only = "" if not self._only else " ONLY"
         name = self._table_name if not self._record_id else f"{self._table_name}:{self._record_id}"
-        return f"CREATE{only} {name}"
+        return f"DELETE{only} {name}"
+
+    def only(self) -> "Delete":
+        self._only = True
+        return self
