+++ src/surrealist/ql/statements/insert.py
@@ -2,8 +2,8 @@ import json
 from typing import List, Dict
 
 from surrealist import Connection
-from surrealist.ql.insert_statements import InsertUseDuplicate
-from surrealist.ql.statement import Statement
+from surrealist.ql.statements.insert_statements import InsertUseDuplicate
+from surrealist.ql.statements.statement import Statement
 from surrealist.utils import OK
